pyproject.toml
[project]
name = "chatbot-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "gradio>=5.42.0",
    "haystack-ai>=2.17.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "docling-haystack",
    "qdrant-client>=1.9.0",
    "docling>=2.45.0",
    "qdrant-haystack>=9.2.0",
    "openai>=1.100.1",
]
